Hid trademark
From my understand of the terms and conditions I don't have to use Spotify Trademarks so for the moment i've commented out the code. Will do more research on this. Also Playing songs from the feed is possible and fixed a bug where the audio player would show when returning from the share controller